Use the value attribute with Radio Buttons and Checkboxes
https://www.freecodecamp.org/learn/responsive-web-design/basic-html-and-html5/use-the-value-attribute-with-radio-buttons-and-checkboxes
Give each of the radio and checkbox inputs the value attribute. Use the input label text, in lowercase, as the value for the attribute.

!!!
If you omit the value attribute, the submitted form data uses the default value, which is on. In this scenario, if the user clicked the "indoor" option and submitted the form, the resulting form data would be indoor-outdoor=on, which is not useful. So the value attribute needs to be set to something to identify the option.